Changement de ces classes

+++ src/classes/Bibliotheque.java
@@ -12,16 +12,16 @@ public class Bibliotheque implements Emprunter
 	CD unCD;
 	Livre unLivre;
 	
-	public void creerUnDVD(String titreMedia, String realAutInt, Date dateParution, double prixHT, int dureeLocation, int uneDuree) {
+	public void creerUnDVD(String titreMedia, String realAutInt, Date dateParution, float prixHT, int dureeLocation, int uneDuree) {
 		unDVD = new DVD(titreMedia, realAutInt, dateParution, prixHT, dureeLocation, uneDuree);
 	}
 	
-	public void creerUnCD(String titreMedia, String realAutInt, Date dateParution, double prixHT, int dureeLocation, int unNbChanson)
+	public void creerUnCD(String titreMedia, String realAutInt, Date dateParution, float prixHT, int dureeLocation, int unNbChanson)
 	{
 		unCD = new CD(titreMedia, realAutInt, dateParution, prixHT, dureeLocation, unNbChanson);
 	}
 	
-	public void creerUnLivre(String titreMedia, String realAutInt, Date dateParution, double prixHT, int dureeLocation, int unNbPages)
+	public void creerUnLivre(String titreMedia, String realAutInt, Date dateParution, float prixHT, int dureeLocation, int unNbPages)
 	{
 		unLivre = new Livre(titreMedia, realAutInt, dateParution, prixHT, dureeLocation, unNbPages);
 	}
